acs_lambda_code/requirements.txt
azure-functions
azure-communication-callautomation
